basic mnist download - to do: decopress

+++ include/cdataset_loader.h
@@ -3,4 +3,4 @@
 */
 #include <stdio.h>
 
-int load_mnist(const char* directory, float* X_train, float* y_train, size_t train_size, float* X_test, float* y_test, size_t test_size);
+int load_mnist(const char* directory, float* X_train, int* y_train, size_t train_size, float* X_test, int* y_test, size_t test_size);
